LPD-59495 Move all form components to common folder
