+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/common/components/forms/PermissionsTable.tsx
