+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/categorization/components/CategorizationPermissionsTable.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 
 import PermissionsTable, {
 	IPermissionItem,
-} from '../../components/forms/PermissionsTable';
+} from '../../../common/components/forms/PermissionsTable';
 import {
 	DEFAULT_PERMISSIONS,
 	DEFAULT_ROLES,
