+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/categorization/vocabularies/EditGeneralInfo.tsx
@@ -16,8 +16,8 @@ import {ClayTooltipProvider} from '@clayui/tooltip';
 import {sub} from 'frontend-js-web';
 import React, {useState} from 'react';
 
+import {IPermissionItem} from '../../../common/components/forms/PermissionsTable';
 import {IVocabulary} from '../../../common/types/IVocabulary';
-import {IPermissionItem} from '../../components/forms/PermissionsTable';
 import CategorizationSpaces from '../components/CategorizationSpaces';
 import PermissionsFormGroup from '../components/PermissionsFormGroup';
